Main simulation logic and entity interactions
Main simulation logic and game object structure, with game turns main loop.

Definition of all the main entities that participate in the game simulation, implementation of all the game board cell types, and interaction with the main class using the mediator pattern.

Implementation of all the card types with their associated actions.